+++ src/main/webapp/index.html
@@ -144,11 +144,34 @@
             </ul>
         </div>
         <div class="col-md-4">
+            <p class="chapter-title">Chapter 15 - Pageflow</p>
+            <ul>
+                <li><a href="http://www.manning.com">Manning.com</a></li>
+            </ul>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4">
+            <p class="chapter-title">Chapter 16 - Tables</p>
+            <ul>
+                <li><a href="people-table.html">People table</a></li>
+            </ul>
+            <ul>
+                <li><a href="city-table.html">City table</a></li>
+            </ul>
+        </div>
+        <div class="col-md-4">
+            <p class="chapter-title">Chapter 17 - Datepicker</p>
+        </div>
+        <div class="col-md-4">
+            <p class="chapter-title">Chapter 18 - jQuery Datepicker</p>
+            <ul>
+                <li><a href="date-picker.html">jQuery UI Date Picker</a></li>
+            </ul>
         </div>
     </div>
     <hr>
     <ul>
-        <li><a href="date-picker.html">jQuery UI Date Picker</a></li>
         <li><a href="listening-div.html">Listening To Key Presses</a></li>
     </ul>
 </div>
