+++ src/test/java/swip/ch16table/city/City.java
@@ -1,26 +1,24 @@
-package swip.ch15table.person;
+package swip.ch16table.city;
 
 import static org.apache.commons.lang3.builder.EqualsBuilder.reflectionEquals;
 import static org.apache.commons.lang3.builder.HashCodeBuilder.reflectionHashCode;
 
-public class Person {
+public class City {
     private final int id;
-    private final String firstName;
-    private final String lastName;
-    private final int points;
+    private final String name;
+    private final String stateName;
 
-    public Person(int id, String firstName, String lastName, int points) {
+    public City(int id, String name, String stateName) {
         this.id = id;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.points = points;
+        this.name = name;
+        this.stateName = stateName;
     }
 
     @Override
     public String toString() {
-        return "new Person(\"" + id +
-                "\",\"" + firstName + "\",\"" +
-                lastName + "\"," + points + ")\n";
+        return "new City(" + id +
+                ",\"" + name + "\",\"" +
+            stateName  + "\")\n";
     }
 
     @Override
